requirements.txt
# Core dependencies
pyyaml>=6.0.0
# asyncio-extra>=1.3.2  # Package not found - commenting out

# JIRA integration
jira>=3.5.0

# HTTP client for API calls
httpx>=0.24.0

# SSH and async SSH connections
paramiko>=3.2.0
asyncssh>=2.13.0

# Database connectivity
asyncpg>=0.28.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
python-dateutil>=2.8.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0